Adjust migrate tests
Adjust build configuration

Use stage0 as test double for dotnet sln and dotnet new

Add some part of production parser code, since all the test ACTs are
just passing the command as string.

Add MigrateTestCommand, which wrap the SystemUnderTest

Replace all TestCommand("dotnet").Execute($"migrate global.json") to
MigrateTestCommand, if not, the test will run against stage0 instead of
SystemUnderTest

Remove WithBackwardsCompatibleRuntimes, since stage0 is an older version
and it has old runtime. The reason of using an old version is, Migrate
is only useful to 1.X run time anyway. And the latest CLI does not have
1.X runtime

“Simplify” GetCurrentRuntimeInformation, in CLI it requires build
output. And it is not needed for a library.